Fix allow negative dates with from [-100,-999]
A bug exists where a date like `-100.11.12` will be incorrectly parsed
as the fast path fails and then fallsback to trying to parse the
synthetic input for the fast path.

This code fixes this by having the fast path return none in the advent
that it can't handle the input.

Benchmarks remain unaffected.