Allow text deserialization on incorrect hint
Continuation of the PR: https://github.com/rakaly/jomini/pull/122

This is helpful to convey preference. Like an object may prefer for the
deserializer to visit a numeric type, but in the advent that the value
isn't numeric, we don't want to fail. Instead it should the downstream
visitor implementation that triggers the failure, not the deserializer.