Writing operator switches to object mode. Trim spaces around equal op
If someone is using the writer but doesn't know what the upcoming
container type is (array, object, mixed). There is no way to switch
without peeking. This allows the writer to always start with an array
and then if an operator is encountered (like an equals operator), the
container will be switched to object mode

And if equals is being written, there shouldn't be any spaces around it.