Add is_empty to TokenResolver trait
Return whether TokenResolver::resolve will always return `None`.

By default this returns `false`

This method is not used by jomini itself, but rather targeted at
downstream save file libraries, who accept an application configured
TokenResolver. If the application is not configured for ironman
support, save file parsers can still handle plain text files, so
`is_empty` allows the save parsers to lazily check the validity of a
TokenResolver when the binary format is encountered. Thus, allowing
for better error messages. Instead of "missing field" errors, the save
file libraries can raise a more descriptive "binary file encountered but
tokens are not configured", as only they know if a non-zero amount of
tokens need to be resolved for a successful deserialization.

There's not a way for jomini to know whether an empty TokenResolver
constitutes an error, as the client may only be deserializing data from
keys that are already strings. Or, alternatively, direct token
deserialization is exclusively used.