Expose a space deliminated format token resolver
Every save game implementation has an $env variable that points to a
file, which the build script turns into a very long `match` statement.
This is poor for compile times and doesn't confer much performance
improvement.

The $env token resolvers are being sunset with tests and implementations
expected to pull from a file.