Perf: consolidate overflow at end of f64 / u64 parsing
Instead of checking if the accumulator overflowed at each digit in a
number. Only run the checks at the end once we know we've consumed
enough digits to even cause an overflow.

Benchmarks are widely positive.

- to_u64/7: unchanged
- to_u64/30: 1.67x faster
- to_u64/1444: 1.50x faster
- to_u64/20405029: 1.60x faster
- to_f64/10: 1.24 slower
- to_f64/-1000: unchanged
- to_f64/0.000: 1.25x faster
- to_f64/20405029.125: 1.50x faster

I optimized these functions as they showed up in as hot spots in
deserialization of data that is number heavy.