Expose data within Paradox save file envelopes
Instead of duplicating similar data extraction across: CK3, Vic3,
Imperator, and EU5 -- might as well standardize on a shared API on how
to process the "envelope" of save data.

This is the result of a bug that came out of faulty CK3 metadata
extraction that had already been fixed in vic3.

This feature is behind an "envelope" feature.