3-4x perf: decouple binary parsing from data representation
I don't fully understand it, but this new binary parser and
deserialization are 3-4x faster. It might be due to decoupling the
binary parsing and data representation, as the new enum `TokenKind` is
only 1 byte large and not 24 bytes like `Token`.

New methods added: `read_token` and `next_token` for the binary
`TokenReader`.

Some results:

- EU4: 3.5x faster
- Vic3: 4x faster
- CK3: 3.8x faster

And in an EU5 application benchmark that included deflate +
deserialization, swapping to this implementation resulted in a 33%
latency reduction: from 480ms to 320ms

As an added benefit the binary deserialization for the reader doesn't
need to use raw pointers anymore as TokenKind doesn't have a lifetime
to the TokenReader.